backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-dotenv==1.2.1
openai==2.13.0
httpx==0.28.1
azure-cognitiveservices-speech==1.47.0
pydantic==2.12.5
requests==2.32.5
